requirements.txt
Django==1.9.4
gunicorn==19.4.5
psycopg2==2.5.3
django-extensions==1.6.1
django-tastypie==0.13.3
django-taggit==0.18.1
django-admin-sortable2==0.6.1
Pillow==3.2.0
elasticsearch>=1.0.0,<2.0.0
django-haystack==2.5.dev1
requests==2.11.1
ipython==5.1.0
jupyter==1.0.0